+++ Sources/iQualize/DreamUI/EQReadoutGrid.swift
@@ -17,7 +17,7 @@ struct EQReadoutGrid: View {
         }
         .padding(.horizontal, 12)
         .padding(.vertical, 6)
-        .background(theme.bgCanvas)
+        .background(ContentMaterialBackground())
     }
 
     @ViewBuilder